% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_theming.R
\name{base_theme}
\alias{base_theme}
\title{base default}
\usage{
base_theme()
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
clean up default clutter in ggplots
}
\details{
TBD
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) + geom_line() +
 base_theme() +
 scale_y_continuous(labels = format_axis_labels_short)
}
base_theme()
}
