% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_colors.R
\docType{data}
\name{cbi_base_darkblue}
\alias{cbi_base_darkblue}
\title{cbi_base_black}
\format{
An object of class \code{character} of length 1.
}
\usage{
cbi_base_darkblue
}
\description{
cbi_base_black
}
\keyword{datasets}
