% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_base_theme.R
\name{cbi_base_theme}
\alias{cbi_base_theme}
\title{Remove defaults from ggplot charts}
\usage{
cbi_base_theme()
}
\value{
A theme object catered toward CBI standards
}
\description{
cbi_base_theme() allows users to call a CBI standardized theme to ggplot charts.
}
\details{
This function removes ggplot() default options that clutter chart space to enhance readability.
 Additionally, it calls upon cbi_geom_defaults() to set standard CBI colors to frequently used plot elements.
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) +
 geom_line() +
 cbi_base_theme()
}
cbi_base_theme()
}
