% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_colors.R
\docType{data}
\name{cbi_base_white}
\alias{cbi_base_white}
\title{cbi_base_lightblue}
\format{
An object of class \code{character} of length 1.
}
\usage{
cbi_base_white
}
\description{
cbi_base_lightblue
}
\keyword{datasets}
