% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_colors.R
\docType{data}
\name{cbi_interestingblue}
\alias{cbi_interestingblue}
\title{'}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
cbi_interestingblue

cbi_interestingblue
}
\description{
'

'
}
\keyword{datasets}
