% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_colors.R
\docType{data}
\name{cbi_interestingblue6}
\alias{cbi_interestingblue6}
\title{'}
\format{
An object of class \code{character} of length 1.
}
\usage{
cbi_interestingblue6
}
\description{
'
}
\keyword{datasets}
