% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_palette.R
\docType{data}
\name{cbi_palettes}
\alias{cbi_palettes}
\title{Predefined CBI colors combined into palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
cbi_palettes
}
\description{
A list of CBI colors combined into palettes that are used
for different charts.
}
\keyword{datasets}
