% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{default_cbi_ds_theme}
\alias{default_cbi_ds_theme}
\title{Format ggplot labels with default plot settings}
\usage{
default_cbi_ds_theme()
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
format_axis_labels_short() formats axis sets defaults to plots.
}
\details{
TBD
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) + geom_line() +
 theme_cbi_ds() +
 scale_y_continuous(labels = format_axis_labels_short)
}
default_cbi_ds_theme()
}
