% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_axes.R
\name{format_axis_labels_short}
\alias{format_axis_labels_short}
\title{Format ggplot labels with short notation}
\usage{
format_axis_labels_short(x, gap = " ", sig_digits_small = NULL, prefix = "")
}
\arguments{
\item{x}{The number to format.}

\item{gap}{Optional character to place between the number and the short notation label. Defaults to space (" ").}

\item{sig_digits_small}{Number of decimals to show for numbers less than 1,000. Defaults to NULL, which means show all available.}

\item{prefix}{Optional character(s) to append to the front of the numbers (e.g., prefix = "$" to add a dollar sign)}
}
\value{
A character vector of the same size as the input.
}
\description{
format_axis_labels_short() formats axis labels in short notation for use in ggplot.
}
\details{
This function allows you to easily format the labels in an axis to print using short notation. The current maximum short notation allowed is (T)rillions. Also allows (B)illions, (M)illions, and thousands (k).
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) + geom_line() +
  scale_y_continuous(labels = format_axis_labels_short)
}
format_axis_labels_short_function()
}
