% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_axes.R
\name{format_financial_data}
\alias{format_financial_data}
\title{Format ggplot labels with dollar notation}
\usage{
format_financial_data(x)
}
\arguments{
\item{x}{The number to format.}
}
\value{
N/A
}
\description{
format_financial_data() formats axis labels in dollar notation for use in ggplot.
}
\details{
This function allows you to easily format the labels in an axis to print using dollar notation. The current maximum short notation allowed is (T)rillions. Also allows (B)illions, (M)illions, and thousands (k).
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) + geom_line() +
  scale_y_continuous(labels = format_financial_data)
}
format_financial_data()
}
