% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_palette.R
\name{make_cbi_pal_discrete}
\alias{make_cbi_pal_discrete}
\title{Create a CBI color palette}
\usage{
make_cbi_pal_discrete(n)
}
\arguments{
\item{n}{number of colors to return}
}
\description{
Function takes CBI graph color palette and returns a vector of colors the size of n.
Used in \code{\link{scale_color_cbi}} and \code{\link{scale_fill_cbi}} to make the discrete
color scale
}
