% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_palette.R
\docType{data}
\name{palette_order}
\alias{palette_order}
\title{Palette order}
\format{
An object of class \code{list} of length 1.
}
\usage{
palette_order
}
\description{
Palette order
}
\keyword{datasets}
