% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_scales.R
\name{scale_color_cbi}
\alias{scale_color_cbi}
\alias{scale_fill_cbi}
\title{Scale constructor for CBI colors}
\usage{
scale_color_cbi(palette = NULL, discrete = FALSE, reverse = FALSE, ...)

scale_fill_cbi(palette = NULL, discrete = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{cbi_palettes}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link{discrete_scale}} or
\code{\link{scale_color_gradientn}}, used respectively when discrete is
TRUE or FALSE}
}
\description{
If you get an error about not enough colors, try a different palette (such
as \code{graph}) or set \code{discrete = FALSE}. If you are plotting discrete colors
and need more than 10 colors, then you will have to make your own palette using
\code{\link{make_cbi_pal}}.
}
\section{Functions}{
\itemize{
\item \code{scale_color_cbi()}: For color scales

\item \code{scale_fill_cbi()}: For fill scales

}}
