% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_base_theme.R
\name{cbi_base_theme}
\alias{cbi_base_theme}
\title{Remove defaults from ggplots}
\usage{
cbi_base_theme()
}
\value{
A theme object catered toward CBI standards
}
\description{
cbi_base_theme() removes ggplot() default that result in confused clutter over readability
}
\details{
This is TBD
}
\examples{
\dontrun{
ggplot(mapping = aes(x = x, y = y), data = data) +
 geom_line() +
 cbi_base_theme()
}
cbi_base_theme()
}
