% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_palette.R
\name{make_cbi_pal_discrete}
\alias{make_cbi_pal_discrete}
\title{Create a CBI color palette}
\usage{
make_cbi_pal_discrete(n)
}
\arguments{
\item{n}{how many colors to return}
}
\description{
Function takes a the CBI graph color palette and returns a vector of colors equal to n.
It is used in \code{\link{scale_color_cbi}} and \code{\link{scale_fill_cbi}} to make the discrete
color scale
}
\seealso{
\code{\link{cbi_palettes}}
}
