% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_scales.R
\name{scale_color_cbi}
\alias{scale_color_cbi}
\alias{scale_fill_cbi}
\title{Scale constructor for CBI colors}
\usage{
scale_color_cbi(palette = NULL, discrete = FALSE, reverse = FALSE, ...)

scale_fill_cbi(palette = NULL, discrete = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Char name of palette in \code{cbi_palettes}}

\item{discrete}{Bool flagging whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Further arguments passed to \code{\link{discrete_scale}} or
\code{\link{scale_colour_gradientn}}}
}
\description{
If plotting discrete data that needs more than 10 colors,
then you will have to create your own palette using
\code{\link{make_cbi_pal}}.
}
\section{Functions}{
\itemize{
\item \code{scale_fill_cbi()}: For fill scales

}}
